+++ src/main/java/com/example/eccomerce/service/interfaces/IProductService.java
@@ -7,12 +7,13 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.util.List;
 
 @Service
 public interface IProductService {
     ResponseProductDto createProduct(String name, String description, String brand,
-                                     int price, ECategory category, MultipartFile file) throws IOException;
+                                     BigDecimal price, ECategory category, MultipartFile file) throws IOException;
 
     //ResponseProductDto create(RequestProductCreateDto user);
